src/python/requirements.txt
chromadb==0.4.17
Flask==3.0.0
Flask-Cors==4.0.0
GitPython==3.1.41
langchain==0.0.335
langchain-community==0.0.13
langchain-core==0.1.13
langdetect==1.0.9
langsmith==0.0.83
openai==1.2.4